Add files via upload
Breaking changes: advance/next/prev no longer bound to end()/cend()/rbegin/crbegin(). If you think a use of these functions may go beyond those locations, check using the < or <= operators and do the bound yourself. Iterators do not have access to the end/begin iterators of a hive, so neither do the advance/next/prev functions. The unneeded metadata removal below forces this change but the reduction in memory usage is deemed more important. Largely-unneeded group metadata variable removed, code fixed to suit. Overall performance benefit, particularly for small scalar types. noexcept removed on hive(block_limits) and hive(block_limits, allocator) due to lakos rule. Various code fluff and minor corrections. Fix to is_active() when a group is deallocated from the hive, but then the same pointer address is re-supplied to the hive via the allocator for a new group, and is_active is called with an iterator pointing to an element in the old (deallocated) group.