Corrections for allocator returning smart pointers
Corrections for when the allocator supplied to the container returns smart pointers. Previously multiple areas would not work due to the need for pointer arithmetic, and if the allocator returned, for example, a unique_ptr, the supplied memory would've been deallocated prematurely. This is also fixed in sort(), where a mismatch of pointer types might've caused problems. Minor opt for pointer_cast in debug mode where all pointer types are trivial.